new reslease

+++ Admin_GUI/Model/usermodel.cpp
@@ -5,7 +5,7 @@ UserModel::UserModel(ISqlDatabaseService *databaseService, LinuxUserService *use
     , m_databaseService(databaseService)
     , m_linuxUserService(userService)
 {
-    DataChanged();
+    ListUpdate();
 }
 
 UserModel::~UserModel()
@@ -15,8 +15,10 @@ UserModel::~UserModel()
 
 int UserModel::GetRoleIdByUserId(const QString &userId) const noexcept
 {
-    for (const User &user : qAsConst(m_users))
+
+    for (int i = 0; i < m_model->rowCount(); ++i)
     {
+        const User user = m_model->item(i)->data(Qt::UserRole + 1).value<User>();
         if (userId == user.GetUserId())
         {
             return user.GetUserRole();
@@ -27,30 +29,33 @@ int UserModel::GetRoleIdByUserId(const QString &userId) const noexcept
 
 void UserModel::AddUserToModel(const QString &userId, const QString &userName, const QString &FCS, int role)
 {
-    for (User &user  : m_users)
+    for (int i = 0; i < m_model->rowCount(); ++i)
     {
-        if (userId == user.GetUserId() && userName == user.GetUserName())
+        const User user = m_model->item(i)->data(Qt::UserRole + 1).value<User>();
+        if (userId == user.GetUserId() && userName == user.GetUserName() )
         {
-            user.SetUserData(FCS, role);
-            m_databaseService->AppendUserIntoTable(user);
-            user.SetUserImage(GetUserImageFromRole(user.GetUserRole()));
-            FillModelByList();
-            return;
+            User newUser(userId, userName, FCS, role);
+            newUser.SetUserImage(GetUserImageFromRole(newUser.GetUserRole()));
+            QStandardItem *item = m_model->item(i);
+            item->setData(QVariant::fromValue(newUser), Qt::UserRole + 1);
+            return ;
         }
     }
     qFatal("%s", QString(Q_FUNC_INFO + QStringLiteral(" Неудачная попытка добавить пользователя. Ид пользователя: ") + userId + QStringLiteral(" имя пользователя: ") + userName + QStringLiteral(" не увенчалась успехом ")).toUtf8().constData());
 }
 
-void UserModel::DeleteUser(const QString &userId)
+void UserModel::DeleteUser(const QString &userId, const QString &userName)
 {
-    for (User &user  : m_users)
+    for (int i = 0; i < m_model->rowCount(); ++i)
     {
-        if (user.GetUserId() == userId)
+        User user = m_model->item(i)->data(Qt::UserRole + 1).value<User>();
+        if (user.GetUserId() == userId && user.GetUserName() == userName)
         {
             m_databaseService->RemoveUserIntoTable(user.GetUserRole(), user);
             user.ClearUserData();
             user.SetUserImage(GetUserImageFromRole(user.GetUserRole()));
-            FillModelByList();
+            QStandardItem *item = m_model->item(i);
+            item->setData(QVariant::fromValue(user), Qt::UserRole + 1);
             return;
         }
     }
@@ -61,60 +66,103 @@ QStandardItemModel *UserModel::GetModel() const noexcept
 {
     return m_model;
 }
-
+/**
+ * void UserModel::OpenFlyAdminSmc()
+ * Отркрываем панель управления пользователями Астры
+ */
 void UserModel::OpenFlyAdminSmc()
 {
     m_linuxUserService->OpenFlyAdminSmc();
 }
-
-void UserModel::DataChanged()
+/**
+ * void UserModel::ListUpdate()
+ * Сперва получаем всех пользователей в системе
+ * Затем удаляем из модели пользователей (если она не пуста) которых там нет
+ * Затем добавляем новых пользвателей
+ * Затем запрашиваем пользователей в бд
+ * И потом их объеденяем
+ */
+void UserModel::ListUpdate()
 {
     Log4QtInfo(Q_FUNC_INFO + QStringLiteral(" Заполняем юзеров инфой из бд в соответствии с ид и именем в системе"));
-    const QList<QPair<QString, QString>> namesAndIdsList = m_linuxUserService->GetSystemUsersNamesWithIds();
-    m_users = FillListByUserService(namesAndIdsList);
+    const QList<QPair<QString, QString>> usersNow = m_linuxUserService->GetAllUsersWithIdInSystem();
+    const QList<QPair<QString, QString> > usersThatNotContains = RemoveNotExsistsUsers(usersNow);
+    AppendNewUsers(usersThatNotContains);
     const QList<User> databaseUsers = FillListByDatabaseService();
-    for (User &user : m_users)
+    JoinModelAndDbList(databaseUsers);
+}
+/**
+ * QList<QPair<QString, QString> > UserModel::RemoveNotExsistsUsers(const QList<QPair<QString, QString> > &namesAndIdsList)
+ * Удаляем из модели пользователей которых нет в системе
+ */
+QList<QPair<QString, QString> > UserModel::RemoveNotExsistsUsers(const QList<QPair<QString, QString> > &namesAndIdsList)
+{
+    QList<QPair<QString, QString> > usersThatNotContains = namesAndIdsList;
+    for (int i = m_model->rowCount() - 1; i >= 0; i--)
     {
-        for (const User &databaseUser : databaseUsers)
+        const User user = m_model->item(i)->data(Qt::UserRole + 1).value<User>();
+        bool finden = false;
+        for (int j = namesAndIdsList.count() - 1; j >= 0; j--)
         {
-            if (user.GetUserId() == databaseUser.GetUserId() && user.GetUserName() == databaseUser.GetUserName())
+            const QPair<QString, QString> nameAndId = namesAndIdsList.at(j);
+            if (nameAndId.first == user.GetUserName() && nameAndId.second == user.GetUserId())
             {
-                user.SetUserData(databaseUser.GetUserFCS(), databaseUser.GetUserRole());
-                user.SetUserImage(GetUserImageFromRole(user.GetUserRole()));
+                finden = true;
+                usersThatNotContains.removeAt(j);
+                break;
             }
         }
+        if (!finden)
+        {
+            m_model->removeRow(i);
+        }
     }
-    FillModelByList();
+    return usersThatNotContains;
 }
-
-QList<User> UserModel::FillListByUserService(const QList<QPair<QString, QString>> &namesAndIdsList) const
+/**
+ * Добавляем новых пользователей в модель если таковые остались
+ */
+void UserModel::AppendNewUsers(const QList<QPair<QString, QString>> &namesAndIdsList)
 {
-    Log4QtInfo(Q_FUNC_INFO + QStringLiteral(" Заполняем юзеров системной инфой "));
-    QList<User> usersInSystem;
-    for (const QPair<QString, QString> &userNameAndId : namesAndIdsList)
+    Log4QtInfo(Q_FUNC_INFO + QStringLiteral(" Дозаполняем юзеров которых в базе нет"));
+    for (const QPair<QString, QString> userNameAndId : qAsConst(namesAndIdsList))
     {
-        User user(userNameAndId.second, userNameAndId.first);
-        user.SetUserImage(GetUserImageFromRole(-1));
-        usersInSystem.append(user);
+        QStandardItem *const item = new QStandardItem();
+        const User user(userNameAndId.second, userNameAndId.first);
+        item->setData(QVariant::fromValue(user), Qt::UserRole + 1);
+        m_model->appendRow(item);
     }
-    return  usersInSystem;
 }
-
+/**
+ * Получаем пользователей из бд
+ */
 QList<User> UserModel::FillListByDatabaseService()
 {
     Log4QtInfo(Q_FUNC_INFO + QStringLiteral(" Запрашиваем юзеров из базы данных "));
     return m_databaseService->GetAllUsers();
 }
-
-void UserModel::FillModelByList() noexcept
+/**
+ * Объеденяем юзеров из таблицы и юзеров из системы
+ * Если у них разные поля, заменяем юзера (на юзера с бд с данными) и добавл ему картинку
+ */
+void UserModel::JoinModelAndDbList(const QList<User> &databaseUsers)
 {
-    m_model->clear();
-    Log4QtInfo(Q_FUNC_INFO + QStringLiteral(" Заполняем модель из листа "));
-    for (const User &user : qAsConst(m_users))
+    for (int i = 0; i < m_model->rowCount(); ++i)
     {
-        QStandardItem *const item = new QStandardItem();
-        item->setData(QVariant::fromValue(user), Qt::UserRole + 1);
-        m_model->appendRow(item);
+        QStandardItem *const item = m_model->item(i);
+        User user = item->data(Qt::UserRole + 1).value<User>();
+        for (User databaseUser : databaseUsers)
+        {
+            if (user.GetUserId() == databaseUser.GetUserId() && user.GetUserName() == databaseUser.GetUserName())
+            {
+                if (user.GetUserFCS() != databaseUser.GetUserFCS()  || user.GetUserRole() != databaseUser.GetUserRole())
+                {
+                    databaseUser.SetUserImage(GetUserImageFromRole(databaseUser.GetUserRole()));
+                    item->setData(QVariant::fromValue(databaseUser), Qt::UserRole + 1);
+                    break;
+                }
+            }
+        }
     }
 }
